deduplicate firefox patch
