+++ liwords-ui/src/utils/hooks.tsx
@@ -1,0 +1,30 @@
+import { useEffect } from 'react';
+
+export const useFirefoxPatch = () => {
+  useEffect(() => {
+    // Prevent backspace unless we're in an input element. We don't want to
+    // leave if we're on Firefox.
+
+    const rx = /INPUT|SELECT|TEXTAREA/i;
+    const evtHandler = (e: KeyboardEvent) => {
+      const el = e.target as HTMLElement;
+      if (e.which === 8) {
+        if (
+          !rx.test(el.tagName) ||
+          (el as HTMLInputElement).disabled ||
+          (el as HTMLInputElement).readOnly
+        ) {
+          e.preventDefault();
+        }
+      }
+    };
+
+    document.addEventListener('keydown', evtHandler);
+    document.addEventListener('keypress', evtHandler);
+
+    return () => {
+      document.removeEventListener('keydown', evtHandler);
+      document.removeEventListener('keypress', evtHandler);
+    };
+  }, []);
+};
